+++ Helltooth/src/ht/graphics/terrain/World.cpp
@@ -1,4 +1,5 @@
 #include "World.h"
+
 #include <cstdio>
 #include <cmath>
 #include <cstdlib>
@@ -52,9 +53,9 @@ namespace ht { namespace graphics {
 	}
 
 	double InterpolatedNoise(int i, double x, double y) {
-		int integer_X = x;
+		int integer_X = (int)x;
 		double fractional_X = x - integer_X;
-		int integer_Y = y;
+		int integer_Y = (int)y;
 		double fractional_Y = y - integer_Y;
 
 		double v1 = SmoothedNoise(i, integer_X, integer_Y),
